requirements.txt
accelerate==0.31.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.33.0
anyio==4.4.0
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.0.1
boto3==1.34.136
botocore==1.34.136
botorch==0.12.0
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
emmet-core==0.84.1
evaluate==0.4.3
filelock==3.15.4
flash_attn==2.5.7
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
gpytorch==1.13
greenlet==2.0.2
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==0.23.4
idna==3.7
ilock==1.0.3
jax==0.4.34
jax_cuda12_pjrt==0.4.34
jax_cuda12_plugin==0.4.34
jaxlib==0.4.34
jaxtyping==0.2.19
jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema_specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.1
langchain-core==0.2.1
langchain-huggingface==0.0.2
langchain-text-splitters==0.2.1
langsmith==0.1.82
latexcodec==3.0.0
linear-operator==0.5.3
-e git+ssh://git@github.com/McGill-NLP/llm2vec.git@250292a307428240d801fadd85825464e71c3277#egg=llm2vec
llvmlite==0.43.0
maggma==0.68.6
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.9.0
ml_dtypes==0.4.0
mmh3==4.1.0
mongomock==4.1.2
monty==2024.5.24
mp-api==0.41.2
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
mypy_extensions==1.0.0
networkx==3.3
ninja==1.11.1
numba==0.60.0
numpy==1.26.4
openai==1.31.0
opt_einsum==3.3.0
orjson==3.10.5
packaging==23.2
palettable==3.3.3
pandas==2.2.1
paramiko==3.4.0
peft==0.12.0
pillow==10.3.0
plotly==5.22.0
portalocker==2.10.1
protobuf==4.25.3
psutil==5.9.8
pybtex==0.24.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydash==8.0.1
pymatgen==2024.4.13
pymongo==4.3.3
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.1.2
pyro_api==0.1.2
pyro_ppl==1.9.1
python_dateutil==2.9.0.post0
python_dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.8.8
requests==2.32.3
rpds_py==0.10.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.7
s3transfer==0.10.2
safetensors==0.4.3
scikit-multilearn==0.2.0
scikit_learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentinels==1.0.0
shutup==0.2.0
six==1.16.0
slingpy==0.2.12
smart-open==7.0.4
sniffio==1.3.1
spglib==2.4.0
SQLAlchemy==2.0.31
sshtunnel==0.4.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.4
transformers==4.44.2
typeguard==4.4.1
typing_extensions==4.12.2
typing_inspect==0.9.0
tzdata==2024.1
umap_learn==0.5.6
uncertainties==3.2.1
urllib3==2.2.2
validators==0.28.3
weaviate-client==4.6.4
wrapt==1.16.0
xxhash==3.2.0
yarl==1.9.4
